REfactor compactor to its own files
